+++ src/Commands/PowerPlatform/PowerAutomate/ExportFlow.cs
@@ -1,7 +1,6 @@
 ﻿using PnP.PowerShell.Commands.Attributes;
 using PnP.PowerShell.Commands.Base;
 using PnP.PowerShell.Commands.Base.PipeBinds;
-using PnP.PowerShell.Commands.Model.PowerPlatform;
 using PnP.PowerShell.Commands.Utilities.REST;
 using System.Management.Automation;
 using System.Net.Http;